refactor(packages): flatten the per-platform routing shims into sibling arms
Folder-backed discovery moved `src/` to `processors/` wholesale, which carried
the per-platform re-export shims across intact: `processors/{apple,linux}/mod.rs`
holding a file-level `#![cfg]` plus `pub mod` + `pub use`, with the real
implementation a level down. The shim only ever existed because the old
`export_plugin!` needed a single Rust path per entry. Per-entry `#[cfg]` landed
earlier in this branch, so there is nothing left for it to unify.

Platform variants are now sibling files under `processors/`, each declaring the
same processor id behind its own mutually-exclusive file-level `#![cfg]` — the
id is the unified name, so nothing routes between the arms:

    processors/audio_capture_linux.rs   #![cfg(target_os = "linux")]
    processors/audio_capture_apple.rs   #![cfg(any(macos, ios))]

and the generated root projects that directly:

    #[cfg(any(target_os = "macos", target_os = "ios"))]
    #[path = "../processors/audio_capture_apple.rs"] pub mod audio_capture_apple;
    #[cfg(target_os = "linux")]
    #[path = "../processors/audio_capture_linux.rs"] pub mod audio_capture_linux;

    streamlib_plugin_abi::export_plugin!(
        #[cfg(any(target_os = "macos", target_os = "ios"))]
        crate::audio_capture_apple::AppleAudioCaptureProcessor::Processor,
        #[cfg(target_os = "linux")]
        crate::audio_capture_linux::LinuxAudioCaptureProcessor::Processor,
        ...
    );

Seven packages had a shim: audio (both arms), camera, display, h264, h265,
jpeg, mp4. Support modules that lived beside the processor inside the shim
become sibling arms carrying the same gate (`v4l2_color_linux.rs`,
`color_vui_translate_linux.rs`, `color_resolved_to_core_linux.rs`), and the
three intra-crate `crate::linux::<support>::` paths follow. Every deleted
`pub use` re-export had zero consumers — each named type is referenced only
inside its own file.

`screen-capture` and `webrtc` needed no change: screen-capture is fully parked,
and `webrtc/processors/streaming/` is an ordinary nested support module with no
`#![cfg]`. Nested directory arms stay supported — a directory is a module
boundary, not a platform mechanism — so the two generator fixtures that
exercised directory arms keep their coverage under a non-platform name
(`processors/capture_backends/`) rather than modelling the shim they document
away. `_apple_impl_pending_/` is untouched apart from one path breadcrumb in
`h264/.../mod.rs` that this rename would otherwise have left dangling.

Verified: `cargo run --locked -p xtask -- install-packages` compiled 15/15
distributable packages through the real `streamlib link --engine` +
`streamlib add` path — the leg that actually `rustc`s the generated crate
roots. `cargo test -p streamlib-processor-extract` is green, including
`real_package_reachability`'s per-target assertions over the flattened audio,
camera, display and screen-capture trees.
