+++ docs/learnings/camera-display-e2e-validation.md
@@ -4,7 +4,7 @@
 
 You changed anything in the GPU pipeline (`vulkan_device.rs`,
 `vulkan_buffer.rs`, `vulkan_texture.rs`,
-`packages/camera/processors/linux/camera.rs`,
+`packages/camera/processors/camera_linux.rs`,
 `linux/processors/display.rs`) and need to confirm:
 
 - Pipeline runs end-to-end without OOM or driver errors
