--- packages/audio/processors/linux/mod.rs
@@ -1,10 +1,0 @@
-// Copyright (c) 2025 Jonathan Fontanez
-// SPDX-License-Identifier: BUSL-1.1
-
-#![cfg(target_os = "linux")]
-
-pub mod audio_capture;
-pub mod audio_output;
-
-pub use audio_capture::{LinuxAudioCaptureProcessor, LinuxAudioInputDevice};
-pub use audio_output::{LinuxAudioDevice, LinuxAudioOutputProcessor};
