+++ packages/h264/processors/_apple_impl_pending_/mod.rs
@@ -10,7 +10,7 @@
 //! imports and do not compile against the package's public-SDK-only
 //! surface yet — when Apple support is activated for `@tatolab/h264`,
 //! rewire the imports the same way the Linux side already did
-//! (`packages/h264/processors/linux/encoder.rs` is the reference).
+//! (`packages/h264/processors/encoder_linux.rs` is the reference).
 //!
 //! The file-level `#![cfg(any())]` above gates the whole directory so it
 //! never compiles. See `packages/mp4/processors/_apple_impl_pending_/`
