--- packages/h265/processors/linux/mod.rs
@@ -1,11 +1,0 @@
-// Copyright (c) 2025 Jonathan Fontanez
-// SPDX-License-Identifier: BUSL-1.1
-
-#![cfg(target_os = "linux")]
-
-pub mod color_vui_translate;
-pub mod decoder;
-pub mod encoder;
-
-pub use decoder::H265DecoderProcessor;
-pub use encoder::H265EncoderProcessor;
