--- packages/jpeg/processors/linux/mod.rs
@@ -1,9 +1,0 @@
-// Copyright (c) 2025 Jonathan Fontanez
-// SPDX-License-Identifier: BUSL-1.1
-
-#![cfg(target_os = "linux")]
-
-pub mod color_resolved_to_core;
-pub mod decoder;
-
-pub use decoder::JpegDecoderProcessor;
