--- packages/mp4/processors/linux/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (c) 2025 Jonathan Fontanez
-// SPDX-License-Identifier: BUSL-1.1
-
-#![cfg(target_os = "linux")]
-
-pub mod mp4_writer;
